requirements.txt
construct
pycryptodome